+++ lib/depgraph/parser.go
@@ -2,11 +2,10 @@ package depgraph
 
 import (
 	"fmt"
-	"io/ioutil"
 	"regexp"
 	"strings"
 
-	"github.com/sirupsen/logrus"
+	"go.uber.org/zap"
 
 	"github.com/Helcaraxan/gomod/lib/internal/util"
 	"github.com/Helcaraxan/gomod/lib/modules"
@@ -17,28 +16,27 @@ var depRE = regexp.MustCompile(`^([^@\s]+)@?([^@\s]+)? ([^@\s]+)@([^@\s]+)$`)
 // GetDepGraph will return the dependency graph for the Go module that can be
 // found at the specified path. The 'logger' parameter can be 'nil' which will
 // result in no output or logging information being provided.
-func GetDepGraph(logger *logrus.Logger, path string) (*DepGraph, error) {
-	if logger == nil {
-		logger = logrus.New()
-		logger.SetOutput(ioutil.Discard)
+func GetDepGraph(log *zap.Logger, path string) (*DepGraph, error) {
+	if log == nil {
+		log = zap.NewNop()
 	}
-	logger.Debug("Creating dependency graph.")
+	log.Debug("Creating dependency graph.")
 
-	mainModule, moduleInfo, err := modules.GetDependencies(logger, path)
+	mainModule, moduleInfo, err := modules.GetDependencies(log, path)
 	if err != nil {
 		return nil, err
 	}
 
-	graph := NewGraph(logger, path, mainModule)
+	graph := NewGraph(log, path, mainModule)
 
-	logger.Debug("Retrieving dependency information via 'go mod graph'")
-	rawDeps, _, err := util.RunCommand(logger, path, "go", "mod", "graph")
+	log.Debug("Retrieving dependency information via 'go mod graph'")
+	rawDeps, _, err := util.RunCommand(log, path, "go", "mod", "graph")
 	if err != nil {
 		return nil, err
 	}
 
 	for _, depString := range strings.Split(strings.TrimSpace(string(rawDeps)), "\n") {
-		graph.logger.Debugf("Parsing dependency: %s", depString)
+		graph.log.Debug("Parsing dependency", zap.String("reference", depString))
 		rawDep, ok := graph.parseDependency(depString, moduleInfo)
 		if !ok {
 			continue
@@ -61,24 +59,24 @@ func (g *DepGraph) addDependency(rawDependency *rawDependency) error {
 		if beginDependency = g.AddDependency(rawDependency.beginModule); beginDependency == nil {
 			return fmt.Errorf("could not create dependency based on %+v", rawDependency.beginModule)
 		}
-		g.logger.Debugf("Created new dependency %q: %+v", rawDependency.beginModule.Path, beginDependency)
+		g.log.Debug("Created new dependency.", zap.String("source", rawDependency.beginModule.Path), zap.Any("dependency", beginDependency))
 	}
 	endDependency, ok := g.GetDependency(rawDependency.endModule.Path)
 	if !ok {
 		if endDependency = g.AddDependency(rawDependency.endModule); endDependency == nil {
 			return fmt.Errorf("could not create dependency based on %+v", rawDependency.endModule)
 		}
-		g.logger.Debugf("Created new dependency %q: %+v", rawDependency.endModule.Path, endDependency)
+		g.log.Debug("Created new dependency.", zap.String("target", rawDependency.endModule.Path), zap.Any("dependency", endDependency))
 	}
 
 	if len(beginDependency.SelectedVersion()) != 0 &&
 		beginDependency.Module.Replace == nil &&
 		beginDependency.SelectedVersion() != rawDependency.beginVersion {
-		g.logger.Warnf(
-			"Encountered unexpected version %q for dependency of %q on %q.",
-			rawDependency.beginVersion,
-			rawDependency.begineName,
-			rawDependency.endName,
+		g.log.Warn(
+			"Encountered unexpected version for a dependency.",
+			zap.String("version", rawDependency.beginVersion),
+			zap.String("source", rawDependency.beginName),
+			zap.String("target", rawDependency.endName),
 		)
 	}
 	beginDependency.Successors.Add(&DependencyReference{
@@ -89,12 +87,17 @@ func (g *DepGraph) addDependency(rawDependency *rawDependency) error {
 		Dependency:        beginDependency,
 		VersionConstraint: rawDependency.endVersion,
 	})
-	g.logger.Debugf("Created new dependency from %q to %q with version %q.", beginDependency.Name(), endDependency.Name(), rawDependency.endVersion)
+	g.log.Debug(
+		"Created new dependency.",
+		zap.String("version", rawDependency.endVersion),
+		zap.String("source", beginDependency.Name()),
+		zap.String("target", endDependency.Name()),
+	)
 	return nil
 }
 
 type rawDependency struct {
-	begineName   string
+	beginName    string
 	beginVersion string
 	beginModule  *modules.Module
 	endName      string
@@ -105,7 +108,7 @@ type rawDependency struct {
 func (g *DepGraph) parseDependency(depString string, moduleMap map[string]*modules.Module) (*rawDependency, bool) {
 	depContent := depRE.FindStringSubmatch(depString)
 	if len(depContent) == 0 {
-		g.logger.Warnf("Skipping ill-formed line in 'go mod graph' output: %s", depString)
+		g.log.Warn("Skipping ill-formed line in 'go mod graph' output.", zap.String("line", depString))
 		return nil, false
 	}
 
@@ -115,14 +118,23 @@ func (g *DepGraph) parseDependency(depString string, moduleMap map[string]*modul
 	beginModule := moduleMap[beginName]
 	endModule := moduleMap[endName]
 	if beginModule == nil || endModule == nil {
-		g.logger.Warnf("Encountered a dependency edge starting or ending at an unknown module %q -> %q.", beginName, endName)
+		g.log.Warn(
+			"Encountered a dependency edge starting or ending at an unknown module.",
+			zap.String("source", beginName),
+			zap.String("target", endName),
+		)
 		return nil, false
 	} else if beginVersion != beginModule.Version {
-		g.logger.Debugf("Skipping edge from %q at %q to %q as we are not using that version.", beginName, beginVersion, endName)
+		g.log.Debug(
+			"Skipping edge as we are not using the specified version.",
+			zap.String("source", beginName),
+			zap.String("version", beginVersion),
+			zap.String("target", endName),
+		)
 		return nil, false
 	}
 	return &rawDependency{
-		begineName:   beginName,
+		beginName:    beginName,
 		beginVersion: beginVersion,
 		beginModule:  beginModule,
 		endName:      endName,